fix: port to current Dalamud API and restore green build
- OtterGui: ObjectKind.MountType renamed to Mount
- ObjectKind.Player renamed to Pc across plugin code
- IClientState.LocalPlayer moved to IObjectTable.LocalPlayer
- IChatGui.ChatMessage handler ported to OnHandleableChatMessageDelegate/IHandleableChatMessage
- SoundManager.GetEffectiveVolume(SoundChannel) replaced with Volume[SoundBus.Voice]
- ClientState.TerritoryType now uint; AddonReceiveEventArgs.AtkEventType now enum
- Services/Queue: add missing System usings
- ConfigWindow: add missing System.Threading.Tasks using
- csproj: skip Windows-only dev-plugin deploy target when APPDATA is unset

+++ VoiceMaster/Helper/Data/LuminaHelper.cs
@@ -13,7 +13,7 @@ namespace VoiceMaster.Helper.DataHelper
 {
     public static class LuminaHelper
     {
-        private static ushort TerritoryRow;
+        private static uint TerritoryRow;
         private static TerritoryType? Territory;
 
         public static TerritoryType? GetTerritory()
